bumps the version separates logic from component by creating a custom hook
