+++ tasks.md
@@ -1,6 +1,6 @@
 # High Level Tasks
 
-1. Add new home page, which will show existing/saved games as well as the possibility to create a new one
+1. ~~Add new home page, which will show existing/saved games as well as the possibility to create a new one~~
 2. Change the folder structure to this (https://redux.js.org/tutorials/essentials/part-2-app-structure) ongoing
    ```
    src/
@@ -32,8 +32,10 @@
 3. Fix `ScoreBoard` component
 4. ~~Add the volleyball~~
 5. ~~Create a new team and save it (Japan's team)~~
-6. Separate framework code from the rest for VolleyballComponent
-7. Add volleyball position to the gameState dependencies
+6. ~~Separate framework code from the rest for VolleyballComponent (separation of concerns)~~
+7. Write down MVP requirements (limit the scope and remove bike-shedding)
+8. Add volleyball position to the gameState dependencies
+9. Perform "separation of concerns" refactoring for the MVP components
 
 ## Out of scope
